pyproject.toml
[project]
name = "model-eval-platform"
version = "0.1.0"
description = "A platform for ML model evaluation and error analysis"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]",
    "pandas",
    "python-multipart",
    "celery",
    "redis",
    "sqlalchemy",
    "psycopg2-binary",
    "minio",
    "pydantic-settings",
    "pytest",  # 测试依赖
    "ruff",     # Lint 依赖
    "mlflow==2.7.1",           # MLOps 核心
    "xgboost", 
    "torch",        # Baseline 模型
    "scikit-learn",     # 数据切分与预处理
    "boto3",            # MLflow 连接 MinIO 需要
    "numpy"             # 数值计算
]

[tool.ruff]
line-length = 120
select = ["E", "F", "I"]

# --- 以下是新增的核心修复部分 ---

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["api", "worker", "core"]